rgw/pubsub rename rest files to follow convention
